refactor: enable linux testing by isolating swiftui
