+++ CombineShim.swift
@@ -1,0 +1,9 @@
+#if !canImport(Combine)
+public protocol ObservableObject: AnyObject {}
+
+@propertyWrapper
+public struct Published<Value> {
+    public var wrappedValue: Value
+    public init(wrappedValue: Value) { self.wrappedValue = wrappedValue }
+}
+#endif
