+++ EditorView.swift
@@ -1,3 +1,4 @@
+#if canImport(SwiftUI)
 import SwiftUI
 import AVFoundation
 
@@ -69,24 +70,6 @@ struct EditorView: View {
     }
 }
 
-enum EditorTool: String, CaseIterable {
-    case trim = "Trim"
-    case captions = "Captions"
-    case effects = "Effects"
-    case filters = "Filters"
-    case audio = "Audio"
-    
-    var icon: String {
-        switch self {
-        case .trim: return "scissors"
-        case .captions: return "text.bubble"
-        case .effects: return "wand.and.stars"
-        case .filters: return "camera.filters"
-        case .audio: return "speaker.wave.2"
-        }
-    }
-}
-
 struct VideoPreviewSection: View {
     let project: VideoProject
     @Binding var isPlaying: Bool
@@ -140,7 +123,6 @@ struct VideoPreviewSection: View {
         .padding(.horizontal)
     }
 }
-
 struct TimelineSection: View {
     @Binding var currentTime: Double
     let duration: Double
@@ -532,4 +514,6 @@ struct SecondaryActionButtonStyle: ButtonStyle {
     EditorView()
         .environmentObject(AppState())
 }
+
+#endif
